+++ AutomatedLabDefinition/AutomatedLabDefinition.psm1
@@ -1,6 +1,4 @@
 ﻿#region Internals
-$script:RedHatPackage = New-Object -TypeName System.Collections.Generic.HashSet[string]
-$script:SusePackage = New-Object -TypeName System.Collections.Generic.HashSet[string]
 $unattendedXmlDefaultContent2012 = @'
 <?xml version="1.0" encoding="utf-8"?>
 <unattend xmlns="urn:schemas-microsoft-com:unattend">
@@ -1674,11 +1672,6 @@ function Add-LabIsoImageDefinition
 
     foreach ($iso in $isos)
     {
-        if ($iso.IsOperatingSystem -and $iso.OperatingSystems.OperatingSystemType -contains 'Linux')
-        {
-            Set-LinuxPackage -Package $iso.OperatingSystems[0].LinuxPackageGroup -LinuxType ($iso.OperatingSystems.LinuxType)[0]
-        }
-
         $isosToRemove = $script:lab.Sources.ISOs | Where-Object { $_.Name -eq $iso.Name -or $_.Path -eq $iso.Path }
         foreach ($isoToRemove in $isosToRemove)
         {
@@ -1924,77 +1917,20 @@ function Add-LabMachineDefinition
 
         [string]$ResourceName,
 
-        [switch]$SkipDeployment
+        [switch]$SkipDeployment,
+
+        [string]$AzureRoleSize,
+
+        [string]$TimeZone,
+
+        [string[]]$RhelPackage,
+
+        [string[]]$SusePackage
     )
-    DynamicParam
-    {
-        $RuntimeParameterDictionary = New-Object System.Management.Automation.RuntimeDefinedParameterDictionary
-
-        #Parameter 'AzureRoleSize'
-        $ParameterName = 'AzureRoleSize'
-        $AttributeCollection = New-Object System.Collections.ObjectModel.Collection[System.Attribute]
-        $ParameterAttribute = New-Object System.Management.Automation.ParameterAttribute
-        $AttributeCollection.Add($ParameterAttribute)
-        $defaultLocation = (Get-LabAzureDefaultLocation -ErrorAction SilentlyContinue).Location
-        if ($defaultLocation)
-        {
-            $vmSizes = Get-AzVMSize -Location $defaultLocation -ErrorAction SilentlyContinue | Where-Object -Property Name -notlike *basic* | Sort-Object -Property Name
-            $validateSetValues = $vmSizes | Select-Object -ExpandProperty Name
-            $ValidateSetAttribute = New-Object System.Management.Automation.ValidateSetAttribute($validateSetValues)
-            $AttributeCollection.Add($ValidateSetAttribute)
-        }
-        $RuntimeParameter = New-Object System.Management.Automation.RuntimeDefinedParameter($ParameterName, [string], $AttributeCollection)
-        $RuntimeParameterDictionary.Add($ParameterName, $RuntimeParameter)
-
-        #Parameter 'TimeZone'
-        $ParameterName = 'TimeZone'
-        $AttributeCollection = New-Object System.Collections.ObjectModel.Collection[System.Attribute]
-        $ParameterAttribute = New-Object System.Management.Automation.ParameterAttribute
-        $AttributeCollection.Add($ParameterAttribute)
-        $validateSetValues = ([System.TimeZoneInfo]::GetSystemTimeZones().Id | Sort-Object)
-        $ValidateSetAttribute = New-Object System.Management.Automation.ValidateSetAttribute($validateSetValues)
-        $AttributeCollection.Add($ValidateSetAttribute)
-        $RuntimeParameter = New-Object System.Management.Automation.RuntimeDefinedParameter($ParameterName, [string], $AttributeCollection)
-
-        $RuntimeParameterDictionary.Add($ParameterName, $RuntimeParameter)
-
-        #Parameter 'RhelPackage'
-        $ParameterName = 'RhelPackage'
-        $AttributeCollection = New-Object System.Collections.ObjectModel.Collection[System.Attribute]
-        $ParameterAttribute = New-Object System.Management.Automation.ParameterAttribute
-        $AttributeCollection.Add($ParameterAttribute)
-        if ($script:RedHatPackage.Count -gt 0)
-        {
-            $ValidateSetAttribute = New-Object System.Management.Automation.ValidateSetAttribute(([string[]]$script:RedHatPackage))
-            $AttributeCollection.Add($ValidateSetAttribute)
-        }
-        $RuntimeParameter = New-Object System.Management.Automation.RuntimeDefinedParameter($ParameterName, [string[]], $AttributeCollection)
-
-        $RuntimeParameterDictionary.Add($ParameterName, $RuntimeParameter)
-
-        #Parameter 'SusePackage'
-        $ParameterName = 'SusePackage'
-        $AttributeCollection = New-Object System.Collections.ObjectModel.Collection[System.Attribute]
-        $ParameterAttribute = New-Object System.Management.Automation.ParameterAttribute
-        $AttributeCollection.Add($ParameterAttribute)
-        if ($script:SusePackage.Count -gt 0)
-        {
-            $ValidateSetAttribute = New-Object System.Management.Automation.ValidateSetAttribute(([string[]]$script:SusePackage))
-            $AttributeCollection.Add($ValidateSetAttribute)
-        }
-
-        $RuntimeParameter = New-Object System.Management.Automation.RuntimeDefinedParameter($ParameterName, [string[]], $AttributeCollection)
-
-        return $RuntimeParameterDictionary
-    }
 
     begin
     {
         Write-LogFunctionEntry
-        $AzureRoleSize = $PsBoundParameters['AzureRoleSize']
-        $TimeZone = $PsBoundParameters['TimeZone']
-        $SusePackage = $PsBoundParameters['SusePackage']
-        $RhelPackage = $PsBoundParameters['RhelPackage']
     }
 
     process
@@ -3114,34 +3050,14 @@ function Get-LabPostInstallationActivity
         [Parameter(ParameterSetName = 'CustomRole')]
         [hashtable]$Properties,
 
-        [switch]$DoNotUseCredSsp
+        [switch]$DoNotUseCredSsp,
+
+        [string]$CustomRole
     )
-    DynamicParam
-    {
-        $RuntimeParameterDictionary = New-Object System.Management.Automation.RuntimeDefinedParameterDictionary
-
-        $ParameterName = 'CustomRole'
-        $AttributeCollection = New-Object System.Collections.ObjectModel.Collection[System.Attribute]
-        $ParameterAttribute = New-Object System.Management.Automation.ParameterAttribute
-        $ParameterAttribute.ParameterSetName = 'CustomRole'
-        $AttributeCollection.Add($ParameterAttribute)
-        $arrSet = (Get-ChildItem -Path (Join-Path -Path (Get-LabSourcesLocationInternal -Local) -ChildPath 'CustomRoles' -ErrorAction SilentlyContinue) -Directory -ErrorAction SilentlyContinue).Name
-
-        if ($arrSet)
-        {
-            $ValidateSetAttribute = New-Object System.Management.Automation.ValidateSetAttribute($arrSet)
-            $AttributeCollection.Add($ValidateSetAttribute)
-            $RuntimeParameter = New-Object System.Management.Automation.RuntimeDefinedParameter($ParameterName, [string], $AttributeCollection)
-
-            $RuntimeParameterDictionary.Add($ParameterName, $RuntimeParameter)
-            return $RuntimeParameterDictionary
-        }
-    }
 
     begin
     {
         Write-LogFunctionEntry
-        $CustomRole = $PsBoundParameters['CustomRole']
         $activity = New-Object -TypeName AutomatedLab.PostInstallationActivity
         if (-not $Properties)
         {
@@ -3798,33 +3714,6 @@ function Repair-LabDuplicateIpAddresses
                 $adapter.Ipv4Address.Add($ipAddress)
             }
         }
-    }
-}
-
-function Set-LinuxPackage
-{
-    param
-    (
-        [string[]]
-        $Package,
-
-        [ValidateSet('RedHat', 'Suse')]
-        [string]
-        $LinuxType
-    )
-
-    if ($LinuxType -eq 'RedHat')
-    {
-        foreach ($entry in $Package)
-        {
-            [void] ($script:RedHatPackage.Add($entry))
-        }
-        return
-    }
-
-    foreach ($entry in $Package)
-    {
-        [void] ($script:SusePackage.Add($entry))
     }
 }